refactor: move source runtime helpers into library
